refactor: simplify kml generator
