+++ packages/libfers/src/CMakeLists.txt
@@ -47,6 +47,7 @@ set(LIBFERS_PRIVATE_HEADERS
 	${CMAKE_CURRENT_LIST_DIR}/serial/xml_parser.h
 	${CMAKE_CURRENT_LIST_DIR}/serial/xml_parser_utils.h
 	${CMAKE_CURRENT_LIST_DIR}/serial/xml_serializer_utils.h
+	${CMAKE_CURRENT_LIST_DIR}/serial/kml_generator_utils.h
 	${CMAKE_CURRENT_LIST_DIR}/signal/dsp_filters.h
 	${CMAKE_CURRENT_LIST_DIR}/signal/radar_signal.h
 	${CMAKE_CURRENT_LIST_DIR}/timing/prototype_timing.h
@@ -87,6 +88,7 @@ set(LIBFERS_SOURCES
 	${CMAKE_CURRENT_LIST_DIR}/serial/xml_parser.cpp
 	${CMAKE_CURRENT_LIST_DIR}/serial/xml_parser_utils.cpp
 	${CMAKE_CURRENT_LIST_DIR}/serial/xml_serializer_utils.cpp
+	${CMAKE_CURRENT_LIST_DIR}/serial/kml_generator_utils.cpp
 	${CMAKE_CURRENT_LIST_DIR}/signal/dsp_filters.cpp
 	${CMAKE_CURRENT_LIST_DIR}/signal/radar_signal.cpp
 	${CMAKE_CURRENT_LIST_DIR}/timing/prototype_timing.cpp
